refactor(frontend): extract FileCard component for better maintainability
- Extract FileCard component from FileList for better separation of concerns
- Move file-related UI logic to dedicated component
- Improve component reusability and maintainability
- Keep consistent styling and functionality

Technical changes:
- Create new FileCard.vue component
- Move file card related props, methods and styles
- Update FileList to use new FileCard component
- Maintain proper event handling between components
- Keep dark mode support
- Preserve all existing functionality

Component responsibilities:
FileList:
- Manages file collection display
- Handles loading and error states
- Manages grid layout
- Communicates with file store

FileCard:
- Handles individual file display
- Manages file actions (download, versions)
- Formats file metadata
- Handles card-level UI interactions
