+++ file-uploading-app/resources/js/app.js
@@ -1,8 +1,15 @@
-import "./bootstrap";
-import { createApp, h } from "vue";
+import { initializeServices } from "./bootstrap";
+import { createApp } from "vue";
 import { createPinia } from "pinia";
 import App from "./App.vue";
-// Create and mount app
-const app = createApp(App);
-app.use(createPinia());
-app.mount("#app");
+
+// Initialize all services before creating Vue app
+initializeServices()
+    .then(() => {
+        const app = createApp(App);
+        app.use(createPinia());
+        app.mount("#app");
+    })
+    .catch((error) => {
+        console.error("Failed to initialize application:", error);
+    });
