+++ file-uploading-app/resources/js/bootstrap.js
@@ -2,15 +2,21 @@ import axios from "./utils/axios";
 import { initializeEcho } from "./utils/echo";
 import { initializePusher } from "./utils/pusher";
 
-// Initialize core services
-initializePusher();
+export async function initializeServices() {
+    try {
+        // Initialize services in sequence
+        await initializePusher();
+        const echo = await initializeEcho();
 
-// Initialize Echo and assign to window
-const echo = initializeEcho();
-if (echo) {
-    window.Echo = echo;
-} else {
-    console.warn("Echo initialization failed");
+        if (echo) {
+            window.Echo = echo;
+        } else {
+            console.warn("Echo initialization failed");
+        }
+
+        console.log("🚀 Application services initialized");
+    } catch (error) {
+        console.error("Service initialization failed:", error);
+        throw error;
+    }
 }
-
-console.log("🚀 Application services initialized");
